+++ tests/bot_detection/test_browserscan.py
@@ -1,7 +1,7 @@
-import truedriver as zd
+import truedriver as td
 
 
-async def test_browserscan(browser: zd.Browser) -> None:
+async def test_browserscan(browser: td.Browser) -> None:
     page = await browser.get("https://www.browserscan.net/bot-detection")
 
     # wait for the page to fully load
@@ -14,6 +14,6 @@ async def test_browserscan(browser: zd.Browser) -> None:
     assert (
         element is not None
         and element.parent is not None
-        and isinstance(element.parent.children[-1], zd.Element)
+        and isinstance(element.parent.children[-1], td.Element)
     )
     assert element.parent.children[-1].text == "Normal"