+++ tests/core/test_browser.py
@@ -1,7 +1,7 @@
 import pytest
 from pytest_mock import MockerFixture
 
-import truedriver as zd
+import truedriver as td
 from tests.conftest import CreateBrowser
 
 
@@ -23,13 +23,13 @@ async def test_connection_error_raises_exception_and_logs_stderr(
     assert "Browser stderr" in caplog.text
 
 
-async def test_get_content_gets_html_content(browser: zd.Browser) -> None:
+async def test_get_content_gets_html_content(browser: td.Browser) -> None:
     page = await browser.get("https://example.com")
     content = await page.get_content()
     assert content.lower().startswith("<!doctype html>")
 
 
-async def test_update_target_sets_target_title(browser: zd.Browser) -> None:
+async def test_update_target_sets_target_title(browser: td.Browser) -> None:
     page = await browser.get("https://example.com")
     await page.update_target()
     assert page.target