+++ tests/core/test_tab.py
@@ -3,13 +3,13 @@ from typing import Any
 
 import pytest
 
-import truedriver as zd
+import truedriver as td
 from tests.sample_data import sample_file
 from truedriver.cdp.fetch import RequestStage
 from truedriver.cdp.network import ResourceType
 
 
-async def test_set_user_agent_sets_navigator_values(browser: zd.Browser) -> None:
+async def test_set_user_agent_sets_navigator_values(browser: td.Browser) -> None:
     tab = browser.main_tab
     assert tab is not None
 
@@ -25,7 +25,7 @@ async def test_set_user_agent_sets_navigator_values(browser: zd.Browser) -> None
     assert navigator_platform == "TestPlatform"
 
 
-async def test_set_user_agent_defaults_existing_user_agent(browser: zd.Browser) -> None:
+async def test_set_user_agent_defaults_existing_user_agent(browser: td.Browser) -> None:
     tab = browser.main_tab
     assert tab is not None
 
@@ -39,7 +39,7 @@ async def test_set_user_agent_defaults_existing_user_agent(browser: zd.Browser) 
     assert navigator_language == "testLang"
 
 
-async def test_find_finds_element_by_text(browser: zd.Browser) -> None:
+async def test_find_finds_element_by_text(browser: td.Browser) -> None:
     tab = await browser.get(sample_file("groceries.html"))
 
     result = await tab.find("Apples")
@@ -49,14 +49,14 @@ async def test_find_finds_element_by_text(browser: zd.Browser) -> None:
     assert result.text == "Apples"
 
 
-async def test_find_times_out_if_element_not_found(browser: zd.Browser) -> None:
+async def test_find_times_out_if_element_not_found(browser: td.Browser) -> None:
     tab = await browser.get(sample_file("groceries.html"))
 
     with pytest.raises(asyncio.TimeoutError):
         await tab.find("Clothes", timeout=1)
 
 
-async def test_select(browser: zd.Browser) -> None:
+async def test_select(browser: td.Browser) -> None:
     tab = await browser.get(sample_file("groceries.html"))
 
     result = await tab.select("li[aria-label^='Apples']")
@@ -66,7 +66,7 @@ async def test_select(browser: zd.Browser) -> None:
     assert result.text == "Apples"
 
 
-async def test_xpath(browser: zd.Browser) -> None:
+async def test_xpath(browser: td.Browser) -> None:
     tab = await browser.get(sample_file("groceries.html"))
 
     results = await tab.xpath('//li[@aria-label="Apples (42)"]')
@@ -79,30 +79,30 @@ async def test_xpath(browser: zd.Browser) -> None:
     assert result.text == "Apples"
 
 
-async def test_add_handler_type_event(browser: zd.Browser) -> None:
+async def test_add_handler_type_event(browser: td.Browser) -> None:
     tab = await browser.get(sample_file("groceries.html"))
 
-    async def request_handler_1(event: zd.cdp.network.RequestWillBeSent) -> None:
+    async def request_handler_1(event: td.cdp.network.RequestWillBeSent) -> None:
         pass
 
-    async def request_handler_2(event: zd.cdp.network.RequestWillBeSent) -> None:
+    async def request_handler_2(event: td.cdp.network.RequestWillBeSent) -> None:
         pass
 
     assert len(tab.handlers) == 0
 
-    tab.add_handler(zd.cdp.network.RequestWillBeSent, request_handler_1)
+    tab.add_handler(td.cdp.network.RequestWillBeSent, request_handler_1)
 
-    tab.add_handler(zd.cdp.network.RequestWillBeSent, request_handler_2)
+    tab.add_handler(td.cdp.network.RequestWillBeSent, request_handler_2)
 
     assert len(tab.handlers) == 1
-    assert len(tab.handlers[zd.cdp.network.RequestWillBeSent]) == 2
-    assert tab.handlers[zd.cdp.network.RequestWillBeSent] == [
+    assert len(tab.handlers[td.cdp.network.RequestWillBeSent]) == 2
+    assert tab.handlers[td.cdp.network.RequestWillBeSent] == [
         request_handler_1,
         request_handler_2,
     ]
 
 
-async def test_add_handler_module_event(browser: zd.Browser) -> None:
+async def test_add_handler_module_event(browser: td.Browser) -> None:
     tab = await browser.get(sample_file("groceries.html"))
 
     async def request_handler(event: Any) -> None:
@@ -110,65 +110,65 @@ async def test_add_handler_module_event(browser: zd.Browser) -> None:
 
     assert len(tab.handlers) == 0
 
-    tab.add_handler(zd.cdp.network, request_handler)
+    tab.add_handler(td.cdp.network, request_handler)
 
     assert len(tab.handlers) == 27
 
 
-async def test_remove_handlers(browser: zd.Browser) -> None:
+async def test_remove_handlers(browser: td.Browser) -> None:
     tab = await browser.get(sample_file("groceries.html"))
 
-    async def request_handler(event: zd.cdp.network.RequestWillBeSent) -> None:
+    async def request_handler(event: td.cdp.network.RequestWillBeSent) -> None:
         pass
 
-    tab.add_handler(zd.cdp.network.RequestWillBeSent, request_handler)
+    tab.add_handler(td.cdp.network.RequestWillBeSent, request_handler)
     assert len(tab.handlers) == 1
 
     tab.remove_handlers()
     assert len(tab.handlers) == 0
 
 
-async def test_remove_handlers_specific_event(browser: zd.Browser) -> None:
+async def test_remove_handlers_specific_event(browser: td.Browser) -> None:
     tab = await browser.get(sample_file("groceries.html"))
 
-    async def request_handler(event: zd.cdp.network.RequestWillBeSent) -> None:
+    async def request_handler(event: td.cdp.network.RequestWillBeSent) -> None:
         pass
 
-    tab.add_handler(zd.cdp.network.RequestWillBeSent, request_handler)
+    tab.add_handler(td.cdp.network.RequestWillBeSent, request_handler)
     assert len(tab.handlers) == 1
 
     tab.remove_handlers(
-        zd.cdp.network.RequestWillBeSent,
+        td.cdp.network.RequestWillBeSent,
     )
     assert len(tab.handlers) == 0
 
 
-async def test_remove_specific_handler(browser: zd.Browser) -> None:
+async def test_remove_specific_handler(browser: td.Browser) -> None:
     tab = await browser.get(sample_file("groceries.html"))
 
-    async def request_handler_1(event: zd.cdp.network.RequestWillBeSent) -> None:
+    async def request_handler_1(event: td.cdp.network.RequestWillBeSent) -> None:
         pass
 
-    async def request_handler_2(event: zd.cdp.network.RequestWillBeSent) -> None:
+    async def request_handler_2(event: td.cdp.network.RequestWillBeSent) -> None:
         pass
 
-    tab.add_handler(zd.cdp.network.RequestWillBeSent, request_handler_1)
-    tab.add_handler(zd.cdp.network.RequestWillBeSent, request_handler_2)
+    tab.add_handler(td.cdp.network.RequestWillBeSent, request_handler_1)
+    tab.add_handler(td.cdp.network.RequestWillBeSent, request_handler_2)
     assert len(tab.handlers) == 1
-    assert len(tab.handlers[zd.cdp.network.RequestWillBeSent]) == 2
+    assert len(tab.handlers[td.cdp.network.RequestWillBeSent]) == 2
 
-    tab.remove_handlers(zd.cdp.network.RequestWillBeSent, request_handler_1)
+    tab.remove_handlers(td.cdp.network.RequestWillBeSent, request_handler_1)
     assert len(tab.handlers) == 1
-    assert len(tab.handlers[zd.cdp.network.RequestWillBeSent]) == 1
+    assert len(tab.handlers[td.cdp.network.RequestWillBeSent]) == 1
 
 
-async def test_remove_handlers_without_event(browser: zd.Browser) -> None:
+async def test_remove_handlers_without_event(browser: td.Browser) -> None:
     tab = await browser.get(sample_file("groceries.html"))
 
-    async def request_handler(event: zd.cdp.network.RequestWillBeSent) -> None:
+    async def request_handler(event: td.cdp.network.RequestWillBeSent) -> None:
         pass
 
-    tab.add_handler(zd.cdp.network.RequestWillBeSent, request_handler)
+    tab.add_handler(td.cdp.network.RequestWillBeSent, request_handler)
     assert len(tab.handlers) == 1
 
     with pytest.raises(ValueError) as e:
@@ -176,7 +176,7 @@ async def test_remove_handlers_without_event(browser: zd.Browser) -> None:
         assert str(e) == "if handler is provided, event_type should be provided as well"
 
 
-async def test_wait_for_ready_state(browser: zd.Browser) -> None:
+async def test_wait_for_ready_state(browser: td.Browser) -> None:
     tab = await browser.get(sample_file("groceries.html"))
 
     await tab.wait_for_ready_state("complete")
@@ -185,15 +185,15 @@ async def test_wait_for_ready_state(browser: zd.Browser) -> None:
     assert ready_state == "complete"
 
 
-async def test_expect_request(browser: zd.Browser) -> None:
+async def test_expect_request(browser: td.Browser) -> None:
     tab = browser.main_tab
     assert tab is not None
 
     async with tab.expect_request(sample_file("groceries.html")) as request_info:
         await tab.get(sample_file("groceries.html"))
         req = await asyncio.wait_for(request_info.value, timeout=3)
-        assert type(req) is zd.cdp.network.RequestWillBeSent
-        assert type(req.request) is zd.cdp.network.Request
+        assert type(req) is td.cdp.network.RequestWillBeSent
+        assert type(req.request) is td.cdp.network.Request
         assert req.request.url == sample_file("groceries.html")
         assert req.request_id is not None
 
@@ -202,15 +202,15 @@ async def test_expect_request(browser: zd.Browser) -> None:
         assert type(response_body) is tuple
 
 
-async def test_expect_response(browser: zd.Browser) -> None:
+async def test_expect_response(browser: td.Browser) -> None:
     tab = browser.main_tab
     assert tab is not None
 
     async with tab.expect_response(sample_file("groceries.html")) as response_info:
         await tab.get(sample_file("groceries.html"))
         resp = await asyncio.wait_for(response_info.value, timeout=3)
-        assert type(resp) is zd.cdp.network.ResponseReceived
-        assert type(resp.response) is zd.cdp.network.Response
+        assert type(resp) is td.cdp.network.ResponseReceived
+        assert type(resp.response) is td.cdp.network.Response
         assert resp.request_id is not None
 
         response_body = await response_info.response_body
@@ -218,7 +218,7 @@ async def test_expect_response(browser: zd.Browser) -> None:
         assert type(response_body) is tuple
 
 
-async def test_expect_download(browser: zd.Browser) -> None:
+async def test_expect_download(browser: td.Browser) -> None:
     tab = browser.main_tab
     assert tab is not None
 
@@ -226,11 +226,11 @@ async def test_expect_download(browser: zd.Browser) -> None:
         await tab.get(sample_file("groceries.html"))
         await (await tab.select("#download_file")).click()
         download = await asyncio.wait_for(download_ex.value, timeout=3)
-        assert type(download) is zd.cdp.browser.DownloadWillBegin
+        assert type(download) is td.cdp.browser.DownloadWillBegin
         assert download.url is not None
 
 
-async def test_intercept(browser: zd.Browser) -> None:
+async def test_intercept(browser: td.Browser) -> None:
     tab = browser.main_tab
     assert tab is not None